refactor: move grid operations to network and network api
And now we can remove the grid-api module.
