+++ refinedstorage-common-api/src/main/java/com/refinedmods/refinedstorage/common/api/grid/Grid.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage.common.api.grid;
 
 import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewProvider;
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridOperations;
 import com.refinedmods.refinedstorage.api.network.node.grid.GridWatcher;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 import com.refinedmods.refinedstorage.api.storage.Storage;
