+++ refinedstorage-common-api/src/main/java/com/refinedmods/refinedstorage/common/api/grid/strategy/GridExtractionStrategy.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.api.grid.strategy;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
 
 import org.apiguardian.api.API;
