+++ refinedstorage-common-api/src/main/java/com/refinedmods/refinedstorage/common/api/grid/view/GridResource.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.api.grid.view;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.resource.repository.ResourceRepository;
