+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/WirelessGrid.java
@@ -2,8 +2,8 @@ package com.refinedmods.refinedstorage.common.grid;
 
 import com.refinedmods.refinedstorage.api.autocrafting.preview.Preview;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
-import com.refinedmods.refinedstorage.api.grid.operations.NoopGridOperations;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridOperations;
+import com.refinedmods.refinedstorage.api.network.node.grid.NoopGridOperations;
 import com.refinedmods.refinedstorage.api.network.node.grid.GridWatcher;
 import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManager;
 import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManagerImpl;
