+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/strategy/ClientGridExtractionStrategy.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.grid.strategy;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
 import com.refinedmods.refinedstorage.common.api.grid.strategy.GridExtractionStrategy;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
 import com.refinedmods.refinedstorage.common.support.packet.c2s.C2SPackets;
