+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/strategy/CompositeGridExtractionStrategy.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.grid.strategy;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
 import com.refinedmods.refinedstorage.common.api.grid.strategy.GridExtractionStrategy;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
