+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/support/resource/ItemResourceType.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage.common.support.resource;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperationsImpl;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridOperations;
+import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridOperationsImpl;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 import com.refinedmods.refinedstorage.api.storage.root.RootStorage;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
