+++ refinedstorage-common/src/test/java/com/refinedmods/refinedstorage/common/grid/query/GridQueryParserImplTest.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.grid.query;
 
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.resource.list.MutableResourceListImpl;
