+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/grid/strategy/FluidGridExtractionStrategy.java
@@ -1,8 +1,8 @@
 package com.refinedmods.refinedstorage.fabric.grid.strategy;
 
 import com.refinedmods.refinedstorage.api.core.Action;
-import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridExtractMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridOperations;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 import com.refinedmods.refinedstorage.api.storage.Storage;
 import com.refinedmods.refinedstorage.common.api.grid.Grid;
