+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/grid/strategy/ItemGridInsertionStrategy.java
@@ -1,8 +1,8 @@
 package com.refinedmods.refinedstorage.fabric.grid.strategy;
 
 import com.refinedmods.refinedstorage.api.core.Action;
-import com.refinedmods.refinedstorage.api.grid.operations.GridInsertMode;
-import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridInsertMode;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridOperations;
 import com.refinedmods.refinedstorage.common.api.grid.Grid;
 import com.refinedmods.refinedstorage.common.api.grid.strategy.GridInsertionStrategy;
 import com.refinedmods.refinedstorage.common.support.resource.ItemResource;
