--- refinedstorage-grid-api/build.gradle.kts
@@ -1,31 +1,0 @@
-plugins {
-    id("refinedarchitect.base")
-}
-
-refinedarchitect {
-    testing()
-    mutationTesting()
-    javadoc()
-    publishing {
-        maven = true
-    }
-}
-
-base {
-    archivesName.set("refinedstorage-grid-api")
-}
-
-dependencies {
-    api(libs.apiguardian)
-    api(project(":refinedstorage-resource-api"))
-    api(project(":refinedstorage-core-api"))
-    api(project(":refinedstorage-storage-api"))
-    api(project(":refinedstorage-query-parser"))
-    implementation(libs.slf4j.api)
-    testImplementation(libs.junit.api)
-    testImplementation(libs.junit.params)
-    testImplementation(libs.assertj)
-    testImplementation(libs.mockito)
-    testRuntimeOnly(libs.junit.engine)
-    testRuntimeOnly(libs.slf4j.impl)
-}
