--- refinedstorage-grid-api/src/main/java/com/refinedmods/refinedstorage/api/grid/operations/package-info.java
@@ -1,7 +1,0 @@
-@ParametersAreNonnullByDefault
-@FieldsAndMethodsAreNonnullByDefault
-package com.refinedmods.refinedstorage.api.grid.operations;
-
-import com.refinedmods.refinedstorage.api.core.FieldsAndMethodsAreNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
