--- refinedstorage-grid-api/src/test/java/com/refinedmods/refinedstorage/api/grid/TestResource.java
@@ -1,10 +1,0 @@
-package com.refinedmods.refinedstorage.api.grid;
-
-import com.refinedmods.refinedstorage.api.resource.ResourceKey;
-
-public enum TestResource implements ResourceKey {
-    A,
-    B,
-    C,
-    D
-}
