+++ refinedstorage-network-api/build.gradle.kts
@@ -18,6 +18,5 @@ dependencies {
     api(project(":refinedstorage-core-api"))
     api(project(":refinedstorage-resource-api"))
     api(project(":refinedstorage-storage-api"))
-    api(project(":refinedstorage-grid-api"))
     api(project(":refinedstorage-autocrafting-api"))
 }
