+++ refinedstorage-network-api/src/main/java/com/refinedmods/refinedstorage/api/network/node/grid/GridExtractMode.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.api.grid.operations;
+package com.refinedmods.refinedstorage.api.network.node.grid;
 
 import org.apiguardian.api.API;
