+++ refinedstorage-network-api/src/main/java/com/refinedmods/refinedstorage/api/network/node/grid/GridOperations.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.api.grid.operations;
+package com.refinedmods.refinedstorage.api.network.node.grid;
 
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.ExtractableStorage;
