+++ refinedstorage-network/build.gradle.kts
@@ -19,7 +19,6 @@ dependencies {
     api(project(":refinedstorage-network-api"))
     api(project(":refinedstorage-resource-api"))
     api(project(":refinedstorage-storage-api"))
-    api(project(":refinedstorage-grid-api"))
     implementation(libs.slf4j.api)
     testImplementation(libs.junit.api)
     testImplementation(libs.junit.params)
