+++ settings.gradle.kts
@@ -47,7 +47,6 @@ include("refinedstorage-core-api")
 include("refinedstorage-resource-api")
 include("refinedstorage-storage-api")
 include("refinedstorage-query-parser")
-include("refinedstorage-grid-api")
 include("refinedstorage-autocrafting-api")
 include("refinedstorage-network-api")
 include("refinedstorage-network")
